+++ etc/etc.go
@@ -25,10 +25,13 @@ import (
 )
 
 //--------------------
-// CONST
+// GLOBAL
 //--------------------
 
-var etcRoot = []string{"etc"}
+var (
+	etcRoot   = []string{"etc"}
+	defaulter = stringex.NewDefaulter("etc", true)
+)
 
 //--------------------
 // VALUE
@@ -102,8 +105,7 @@ type Etc interface {
 
 // etc implements the Etc interface.
 type etc struct {
-	values    collections.KeyStringValueTree
-	defaulter stringex.Defaulter
+	values collections.KeyStringValueTree
 }
 
 // Read reads the SML source of the configuration from a
@@ -122,8 +124,7 @@ func Read(source io.Reader) (Etc, error) {
 		return nil, errors.Annotate(err, ErrIllegalSourceFormat, errorMessages)
 	}
 	return &etc{
-		values:    tree,
-		defaulter: stringex.NewDefaulter("etc", true),
+		values: tree,
 	}, nil
 }
 
@@ -146,51 +147,49 @@ func ReadFile(filename string) (Etc, error) {
 // ValueAsString implements the Etc interface.
 func (e *etc) ValueAsString(path, dv string) string {
 	value := e.valueAt(path)
-	return e.defaulter.AsString(value, dv)
+	return defaulter.AsString(value, dv)
 }
 
 // ValueAsBool implements the Etc interface.
 func (e *etc) ValueAsBool(path string, dv bool) bool {
 	value := e.valueAt(path)
-	return e.defaulter.AsBool(value, dv)
+	return defaulter.AsBool(value, dv)
 }
 
 // ValueAsInt implements the Etc interface.
 func (e *etc) ValueAsInt(path string, dv int) int {
 	value := e.valueAt(path)
-	return e.defaulter.AsInt(value, dv)
+	return defaulter.AsInt(value, dv)
 }
 
 // ValueAsFloat64 implements the Etc interface.
 func (e *etc) ValueAsFloat64(path string, dv float64) float64 {
 	value := e.valueAt(path)
-	return e.defaulter.AsFloat64(value, dv)
+	return defaulter.AsFloat64(value, dv)
 }
 
 // ValueAsTime implements the Etc interface.
 func (e *etc) ValueAsTime(path, format string, dv time.Time) time.Time {
 	value := e.valueAt(path)
-	return e.defaulter.AsTime(value, format, dv)
+	return defaulter.AsTime(value, format, dv)
 }
 
 // ValueAsDuration implements the Etc interface.
 func (e *etc) ValueAsDuration(path string, dv time.Duration) time.Duration {
 	value := e.valueAt(path)
-	return e.defaulter.AsDuration(value, dv)
+	return defaulter.AsDuration(value, dv)
 }
 
 // Split implements the Etc interface.
 func (e *etc) Split(path string) (Etc, error) {
-	pathParts := strings.Split(path, "/")
-	fullPath := append(etcRoot, pathParts...)
+	fullPath := makeFullPath(path)
 	values, err := e.values.CopyAt(fullPath...)
 	if err != nil {
 		return nil, errors.Annotate(err, ErrCannotSplit, errorMessages)
 	}
 	values.At(fullPath[len(fullPath)-1:]...).SetKey("etc")
 	es := &etc{
-		values:    values,
-		defaulter: e.defaulter,
+		values: values,
 	}
 	return es, nil
 }
@@ -198,12 +197,11 @@ func (e *etc) Split(path string) (Etc, error) {
 // Apply implements the Etc interface.
 func (e *etc) Apply(appl Application) (Etc, error) {
 	ec := &etc{
-		values:    e.values.Copy(),
-		defaulter: e.defaulter,
+		values: e.values.Copy(),
 	}
-	for key, value := range appl {
-		path := append(etcRoot, strings.Split(key, "/")...)
-		_, err := ec.values.Create(path...).SetValue(value)
+	for path, value := range appl {
+		fullPath := makeFullPath(path)
+		_, err := ec.values.Create(fullPath...).SetValue(value)
 		if err != nil {
 			return nil, errors.Annotate(err, ErrCannotApply, errorMessages)
 		}
@@ -219,7 +217,7 @@ func (e *etc) String() string {
 // valueAt retrieves and encapsulates the value
 // at a given path.
 func (e *etc) valueAt(path string) *value {
-	fullPath := append(etcRoot, strings.Split(path, "/")...)
+	fullPath := makeFullPath(path)
 	changer := e.values.At(fullPath...)
 	return &value{fullPath, changer}
 }
@@ -227,6 +225,11 @@ func (e *etc) valueAt(path string) *value {
 //--------------------
 // HELPERS
 //--------------------
+
+// makeFullPath creates the full path out of a string.
+func makeFullPath(path string) []string {
+	return append(etcRoot, strings.Split(path, "/")...)
+}
 
 // pathToString returns the path in a filesystem like notation.
 func pathToString(path []string) string {